Update Dockerfile to install into /opt + incl. sambamba
When running with Singularity and workflow engines, often we don't get a choice about which user we log in as. I've moved the install files to instead be placed in `/opt` by convention, but link them back to data for backwards compatibility.

I've also installed the latest version of Sambamba as required in GRIDSS >= 2.5.0.